backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
PyJWT
argon2-cffi
watchfiles
